refactor: Parser를 구분할 필요가 없어보여 지우고 InputController에 합침
